requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
onnx-simplifier==0.4.36
onnxruntime-gpu==1.22.0
onnxslim==0.1.67
opencv-python==4.12.0.88
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
polars==1.33.0
protobuf==6.32.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.1
scipy==1.15.3
segment-anything-py==1.0.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
stringzilla==4.4.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.194
ultralytics-thop==2.0.17
urllib3==2.5.0
